Refactor: Move AssemblyError to errors crate to fix leaky abstraction
- Moved `AssemblyError` from `src/semantic/assembler.rs` to `src/errors/assembly.rs`.
- Updated `GlossaError` to include `AssemblyError` as a transparent variant.
- Updated `src/semantic/mod.rs` to propagate `AssemblyError` directly instead of stringifying it.
- Added `miette::Diagnostic` derive to `AssemblyError`.
- Updated `src/errors/mod.rs` to export `assembly` module.
- Updated journal in `.jules/atlas.md`.

This change allows `AssemblyError` to be propagated with full structural information, improving error reporting and diagnostics, while maintaining acyclic dependencies.
